Move utils.restingstate to interfaces.resting_state.
