+++ xcp_d/interfaces/__init__.py
@@ -9,7 +9,14 @@ from xcp_d.interfaces.qc_plot import QCPlot
 from xcp_d.interfaces.regression import CiftiDespike, Regress
 from xcp_d.interfaces.report import AboutSummary, FunctionalSummary, SubjectSummary
 from xcp_d.interfaces.report_core import generate_reports
-from xcp_d.interfaces.resting_state import BrainPlot, ComputeALFF, SurfaceReHo
+from xcp_d.interfaces.resting_state import (
+    BrainPlot,
+    ComputeALFF,
+    ContrastEnhancement,
+    DespikePatch,
+    ReHoNamePatch,
+    SurfaceReHo,
+)
 from xcp_d.interfaces.surfplotting import (
     BrainPlotx,
     PlotImage,
@@ -22,6 +29,7 @@ from xcp_d.interfaces.surfplotting import (
 __all__ = [
     'Regress', 'FilteringData', 'NiftiConnect',
     'ComputeALFF', 'SurfaceReHo',
+    'ReHoNamePatch', 'DespikePatch', 'ContrastEnhancement',
     'ApplyTransformsx', 'Interpolate', 'CensorScrub', 'RemoveTR',
     'QCPlot', 'SubjectSummary', 'AboutSummary', 'FunctionalSummary',
     'generate_reports', 'CiftiDespike', 'ConnectPlot', 'BrainPlot',
